third_party/WebKit/LayoutTests/fast/dom/Element/matches.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body class="foo">
<script>
var element = document.body;
["matches", "webkitMatchesSelector"].forEach(function(func)
{
    shouldThrow("element." + func + "()");
    shouldBeTrue("element." + func + "('body')");
    shouldBeTrue("element." + func + "('html > .foo')");
    shouldBeFalse("element." + func + "('html')");
    shouldBeFalse("element." + func + "('html > .bar')");
    shouldThrow("element." + func + "('')");
    shouldThrow("element." + func + "('!invalid!')");
});
</script>
</body>
</html>
